% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_distance}
\alias{cloud_distance}
\title{cloud_distance computes the distance between two cloud objects}
\usage{
cloud_distance(x, y)
}
\arguments{
\item{x, y}{cloud objects}
}
\value{
numeric, the distance between clouds
}
\description{
cloud_distance computes the distance between two cloud objects
}
