% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{j_clouds}
\alias{j_clouds}
\title{j-clouds clustering}
\usage{
j_clouds(df, var = NULL, group = NULL, k = NULL)
}
\arguments{
\item{df}{data frame with the data to cluster}

\item{var}{character or integer indicating the column name or position
of the numeric column to evalute for defining the clusters, 1L
by default}

\item{group}{character or integer indicating the column name or position
of the column that defines the treatments, 2L by default}

\item{k}{integer, number of super-groups to cluster in (where to cut
the hierarchichal tree)}
}
\value{
a data frame with columns group and cluster.
                - group: referencies the group variable
                - cluster: a letter indicating the cluster to which the
                           group belongs
}
\description{
j-clouds clustering
}
