% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{k_clouds}
\alias{k_clouds}
\title{k-clouds clustering}
\usage{
k_clouds(df, var = NULL, group = NULL, k = NULL, threshold = 0.01,
  iter = 100)
}
\arguments{
\item{df}{data frame with the data to cluster}

\item{var}{character or integer indicating the column name or position
of the numeric column to evalute for defining the clusters,
1L by default}

\item{group}{character or integer indicating the column name or position
of the column that defines the treatments, 2L by default}

\item{k}{integer, number of super-groups to cluster in}

\item{threshold}{double, value to stop the iteration. Stops if the
difference between the previous total distance and the
current total distance is less than threshold}

\item{iter}{integer, maximum number of iterations}
}
\value{
a data frame with columns group and cluster.
                    - group: referencies the group variable
                    - cluster: a letter indicating the cluster to which the
                               group belongs
}
\description{
k-clouds clustering
}
